requirements.txt
numpy==1.19.0
Pillow==7.1.2
torch==1.5.1
torchvision==0.6.1
argparse==1.4.0
opencv-python==4.2.0.34
